Split pulse railway stack module
